requirements.txt
certifi==2018.1.18
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==2.0.3
django-cors-headers==2.2.0
django-heroku==0.3.1
django-money==0.12.3
django-rest-framework==0.1.0
djangorestframework==3.7.7
gunicorn==19.7.1
idna==2.6
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.0
psycopg2==2.7.4
py-moneyed==0.7.0
pytz==2018.3
requests==2.18.4
uritemplate==3.0.0
urllib3==1.22
whitenoise==3.3.1
pillow==5.2.0
